perf: :zap: Changed some of the functions of the endpoints and moved to a routes file
Now the endpoints work on a different file and they were improved in performance
